+++ src/D2NG/BNCS/Packet/AuthInfoRequestPacket.cs
@@ -35,7 +35,7 @@ namespace D2NG.BNCS.Packet
         public AuthInfoRequestPacket(int version)
             : base(
                 BuildPacket(
-                    (byte)Sid.AUTH_INFO,
+                    Sid.AUTH_INFO,
                     ProtocolId,
                     Encoding.ASCII.GetBytes(PlatformCode).Reverse().ToArray(),
                     Encoding.ASCII.GetBytes(ProductCode).Reverse().ToArray(),
