+++ src/D2NG/BNCS/Packet/LogonRequestPacket.cs
@@ -13,7 +13,7 @@ namespace D2NG.BNCS.Packet
             string password) :
             base(
                 BuildPacket(
-                    (byte)Sid.LOGONRESPONSE2,
+                    Sid.LOGONRESPONSE2,
                     BitConverter.GetBytes(clientToken),
                     BitConverter.GetBytes(serverToken),
                     Bsha1.DoubleHash(clientToken, serverToken, password.ToLowerInvariant()),
